--- a/transport/http2_client.go
+++ b/transport/http2_client.go
@@ -409,13 +409,9 @@ func (t *http2Client) NewStream(ctx context.Context, callHdr *CallHdr) (_ *Strea
 		return nil, ErrConnClosing
 	}
 	t.mu.Unlock()
-	sq, err := wait(ctx, t.ctx, nil, nil, t.streamsQuota.acquire())
-	if err != nil {
+	// Get a quota of 1 from streamsQuota.
+	if _, _, err := t.streamsQuota.get(1, waiters{ctx: ctx, tctx: t.ctx}); err != nil {
 		return nil, err
-	}
-	// Returns the quota balance back.
-	if sq > 1 {
-		t.streamsQuota.add(sq - 1)
 	}
 	// TODO(mmukhi): Benchmark if the performance gets better if count the metadata and other header fields
 	// first and create a slice of that exact size.
@@ -664,7 +660,6 @@ func (t *http2Client) Write(s *Stream, hdr []byte, data []byte, opts *Options) e
 		streamQuotaVer uint32
 		localSendQuota int
 		err            error
-		sqChan         <-chan int
 	)
 	for idx, r := range [][]byte{hdr, data} {
 		for len(r) > 0 {
@@ -673,16 +668,25 @@ func (t *http2Client) Write(s *Stream, hdr []byte, data []byte, opts *Options) e
 				size = len(r)
 			}
 			if streamQuota == 0 { // Used up all the locally cached stream quota.
-				sqChan, streamQuotaVer = s.sendQuotaPool.acquireWithVersion()
-				// Wait until the stream has some quota to send the data.
-				streamQuota, err = wait(s.ctx, t.ctx, s.done, s.goAway, sqChan)
+				// Get all the stream quota there is.
+				streamQuota, streamQuotaVer, err = s.sendQuotaPool.get(math.MaxInt32, waiters{
+					ctx:    s.ctx,
+					tctx:   t.ctx,
+					done:   s.done,
+					goAway: s.goAway,
+				})
 				if err != nil {
 					return err
 				}
 			}
 			if localSendQuota <= 0 { // Being a soft limit, it can go negative.
 				// Acquire local send quota to be able to write to the controlBuf.
-				localSendQuota, err = wait(s.ctx, t.ctx, s.done, s.goAway, s.localSendQuota.acquire())
+				localSendQuota, _, err = s.localSendQuota.get(math.MaxInt32, waiters{
+					ctx:    s.ctx,
+					tctx:   t.ctx,
+					done:   s.done,
+					goAway: s.goAway,
+				})
 				if err != nil {
 					return err
 				}
@@ -690,16 +694,18 @@ func (t *http2Client) Write(s *Stream, hdr []byte, data []byte, opts *Options) e
 			if size > streamQuota {
 				size = streamQuota
 			} // No need to do that for localSendQuota since that's only a soft limit.
-			// Wait until the transport has some quota to send the data.
-			tq, err := wait(s.ctx, t.ctx, s.done, s.goAway, t.sendQuotaPool.acquire())
+			// Get size worth quota from transport.
+			tq, _, err := t.sendQuotaPool.get(size, waiters{
+				ctx:    s.ctx,
+				tctx:   t.ctx,
+				done:   s.done,
+				goAway: s.goAway,
+			})
 			if err != nil {
 				return err
 			}
 			if tq < size {
 				size = tq
-			}
-			if tq > size { // Overbooked transport quota. Return it back.
-				t.sendQuotaPool.add(tq - size)
 			}
 			streamQuota -= size
 			localSendQuota -= size