In the fast path instead of both acquiring a lock and writing on a channel, quota pool now only acquires the lock. In the slow path however, it needs to do both. 

Following are benchmark results:

### gRPC-Go benchmakrs:
Ran streaming ping pong benchmarks with 1000 concurrent streams and varied request-response sizes.

#### Stream-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_1000-reqSize_1B-respSize_1B-Compressor_false: 

Before
```
50_Latency: 5852.5600 µs | 90_Latency: 8644.2170 µs | 99_Latency: 12326.9280 µs | Avg latency: 6096.6670 µs | Count: 9825882 | 835 Bytes/op | 31 Allocs/op |  
```

After
```
50_Latency: 2065.5250 µs | 90_Latency: 2782.8890 µs | 99_Latency: 5785.2260 µs | Avg latency: 2212.1370 µs | Count: 27098481 | 825 Bytes/op | 30 Allocs/op |  
```


#### Stream-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_1000-reqSize_100B-respSize_100B-Compressor_false:

Before
```
50_Latency: 6383.2990 µs | 90_Latency: 9348.2500 µs | 99_Latency: 13037.6390 µs | Avg latency: 6483.3020 µs | Count: 9242647 | 2490 Bytes/op | 31 Allocs/op |  
```

After
```
50_Latency: 2520.5400 µs | 90_Latency: 3535.2930 µs | 99_Latency: 7008.4330 µs | Avg latency: 2749.8300 µs | Count: 21809014 | 2475 Bytes/op | 30 Allocs/op
```


####  Stream-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_1000-reqSize_1024B-respSize_1024B-Compressor_false:

Before
```
50_Latency: 8374.0940 µs | 90_Latency: 12119.2840 µs | 99_Latency: 17004.9960 µs | Avg latency: 8447.3220 µs | Count: 7091109 | 18516 Bytes/op | 31 Allocs/op
```

After
```
50_Latency: 5253.6050 µs | 90_Latency: 10363.1550 µs | 99_Latency: 18257.3020 µs | Avg latency: 6233.3960 µs | Count: 9613785 | 18231 Bytes/op | 30 Allocs/op
```

#### Stream-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_1000-reqSize_65536B-respSize_65536B-Compressor_false:

Before
```
50_Latency: 103.2264 ms | 90_Latency: 148.4969 ms | 99_Latency: 168.8514 ms | Avg latency: 107.0320 ms | Count: 560864 | 969070 Bytes/op | 67 Allocs/op
```

After
```
50_Latency: 104503.3340 µs | 90_Latency: 147591.8060 µs | 99_Latency: 194919.4060 µs | Avg latency: 100270.9460 µs | Count: 598807 | 938790 Bytes/op | 73 Allocs/op
```

#### Stream-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_1000-reqSize_1048576B-respSize_1048576B-Compressor_false: 

Before
```
50_Latency: 1748.3104 ms | 90_Latency: 2404.2216 ms | 99_Latency: 6274.2200 ms | Avg latency: 1863.8317 ms | Count: 32668 | 14114507 Bytes/op | 646 Allocs/op
```

After
```
50_Latency: 1455.0151 ms | 90_Latency: 1724.1157 ms | 99_Latency: 4061.2610 ms | Avg latency: 1498.5305 ms | Count: 40510 | 13723061 Bytes/op | 717 Allocs/op
```

This also shows significant improvement on [pinger benchmark](https://github.com/petermattis/pinger) by @petermattis 

Before
```
 ./pinger -c 200 -p 1000 -n 1 -d 10s -t grpc localhost:50051
_elapsed____ops/s_____MB/s__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
     10s 102770.7     39.2     13.1     15.7     18.9     37.7

```

After
```
 ./pinger -c 200 -p 1000 -n 1 -d 10s -t grpc localhost:50051
_elapsed____ops/s_____MB/s__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
     10s 177063.1     67.5      0.0     14.7     18.9     46.1
```


In relation to #1512