@petermattis 2017-11-01
Nice! I was wondering if avoiding the channel operations for quota acquisition was possible.